Unification of UI, obscuring FAB button fix
This commit introduces changes to UI - now all buttons are ElevatedButtons, also the elevation is reduced to 8dp. ShoppingItemEditor is now Card object. Added padding at the bottom of Items list, so FAB button does not obscure items at the end of the list preventing access to "edit" and "delete" buttons.